requirements.txt
pandas
pytube
yt_dlp
pydub
langchain
langchain-community
langchain-openai
langchain-chroma
openai
python-dotenv
tqdm
chromadb
tiktoken
youtube-transcript-api